refactor: API UserQueryAdminController로 이동 (#1190)
